Remove speed in live activity and turn average speed into Int

+++ BougeTesJambes/Model/SessionModel.swift
@@ -13,7 +13,7 @@ struct SessionModel: Identifiable, Hashable, Codable {
 
 	let sessionTime: Int
 	let sessionDistanceInMeters: Double
-	let sessionAverageSpeed: Double
+	let sessionAverageSpeed: Int
 	let pace: Int
 	let distanceSpeedChart: [DistanceSpeedChart]?
 	let timeSpeedChart: [TimeSpeedChart]?
